Factor out search strategies into its own file
